Refactored TCP Interface
